sc_event::none - convert from global variable to function
The sc_event::none variable was introduced in
55da81d96e73ea8bccad1dbab59c1bae92cee661. The problem with the
introduced implementation is that the created event contains a
reference to the simulation context. This simulation context is created
by the sc_event constructor through the invocation of
sc_get_curr_simcontext() on program startup.

Previously, it was possible to reset the simulation context by setting
the global sc_curr_simcontext variable to NULL after each invocation of
sc_elab_and_sim(). This caused sc_get_curr_simcontext() to allocate and
return a new context on the first invocation thereafter. However, with
the introduced sc_event::none change, the none event will always hold a
reference to the initial (possibly freed) sc_curr_simcontext.

This commit proposes a change which tracks the none event as part of the
simulation context. The change converts sc_event::none to a static
member function which returns a reference to the m_none member of the
current simulation context.

Neither sc_event::none nor sc_get_curr_simcontext() are currently
mandated by IEEE Std 1666-2011. For this reason, this change should not
affected existing standard-conforming SystemC programs.

Signed-off-by: Sören Tempel <tempel@uni-bremen.de>