sc_process_handle::non_event - convert from global variable to function
With the previous commit applied, sc_process_handle::non_event is (to
the best of my current knowledge) the last remaining sc_event declared
as a global variable. As the sc_event class constructor obtains a
reference to the global simulation context via sc_get_curr_simcontext()
having global sc_events causes the simulation context to be initialized
at program startup. With this commit applied, the simulation context is
only initialized when needed.

This commit couples the sc_process_handle::non_event with the simulation
context. This allows proper resets of the simulation context by
assigning it to NULL, causing the next invocation of
sc_get_curr_simcontext() thereafter to allocate a new context. Without
this commit, the sc_process_handle::non_event would contain a reference
to a freed simulation context when doing so.

Signed-off-by: Sören Tempel <tempel@uni-bremen.de>